Some cleaning up, triggered by alpha.
- TARGET_CPU_DEFAULT belongs in configure.in, not in various .h files.
If it goes out wrong, it's the output of uname that needs to be fixed
(or config.guess, e.g., even pentium gives out i386, as intended). Makes
it possible to configure a i586-unknown-openbsd system if one really WANTS
to output only pentium specific code.

- ASM_FILE_START put into alpha/openbsd.h. Used to be in alpha.h, but
elf & osf diverged...

- set xfrag=none for alpha, as the default config is bogus and supplies
a dec-osf specific xalpha.