Fix typo in xemacs installation directory.
This should be the final solution to that problem.
* basic package needs BOTH emacs and xemacs to build, as it can't rely
on emacs-lisp compilers being intercheangable.
* one can build the port without emacs or xemacs, by asking explicitly for
it.
* basic package installs TWO sets of emacs files, one for emacs, one for
xemacs.

There needn't be RUNNING_DEPENDENCIES between id-utils and emacs/xemacs,
as the emacs-lisp files are an add-on, not necessary for id-utils to work.

However, pkg_* don't currently support the notion of a `shared' directory
between several packages, hence a succession of

pkg_add emacs
pkg_add id-utils
pkg_delete emacs
pkg_delete id-utils

won't deal with /usr/local/share/emacs/site_lisp correctly...
This is a problem in pkg_* design, not in the id-utils port.

Package checked and uploaded.