Update to version 19991231-pl04. This fix is urgent.
-- Info from Wietse --
Patch 04 fixes one old problem that never happened, and fixes two
new problems introduced with patch 03, one of which urgent.  The
problems were found as I was reorganizing the cleanup daemon code.

The too long header fix in patch 03 loses mail when a header is
found too long, and when that mail is't posted with "sendmail -t".

Actually the mail is not lost. The queue file stays in the incoming
queue, as a file with mode -rw-------.

In order to have such mail delivered, chdir into the directory
/var/spool/postfix/incoming, and look for mode -rw------- queue
files that are older than a few minutes (ls -lt|more) and chmod +x
those old mode -rw------- queue files.

DO NOT CHMOD +X ALL INCOMING QUEUE FILES because that causes Postfix
to discard messages that are still being received.