Dependency tweaks.
- put .if defined(XXX) && !defined(YYY) on the same line: less indentation,
more readability.
- set found to true/false directly (builtin).

ALWAYS check that a dependency directory does exist and is a directory.
Catch up typos more quickly, since they will be evident *even if the
dependency is already satisfied*.

Make this only a warning, not a hard error, as we still want to cater
to users with a partially checked out tree...