import libwww-5.48 from CPAN ok by brad@ --
Libwww-perl is a collection of Perl modules which provides a simple
and consistent application programming interface (API) to the
World-Wide Web.

The main focus of the library is to provide classes and functions
that allow you to write WWW clients, thus libwww-perl said to be a
WWW client library. The library also contain modules that are of
more general use and even classes that help you implement simple
HTTP servers.