- import of p5-XML-Parser-2.30
--
This module provides ways to parse XML documents. It is built on
top of XML::Parser::Expat, which is a lower level interface to James
Clark's expat library. Each call to one of the parsing methods
creates a new instance of XML::Parser::Expat which is then used to
parse the document. Expat options may be provided when the XML::Parser
object is created.

These options are then passed on to the Expat object on each parse
call. They can also be given as extra arguments to the parse methods,
in which case they override options given at XML::Parser creation
time.