[ security update ]
update to p5-libwww-5.51

If LWP::UserAgent::env_proxy is called in a CGI environment, the
case-insensitivity when looking for "http_proxy"
permits "HTTP_PROXY" to be found, but this can be trivially set by the
web client using the "Proxy:" header.  The fix applied is that
$ENV{HTTP_PROXY} is not longer honored for CGI scripts.
The CGI_HTTP_PROXY environment variable can be used instead.