Import of fvwm-2.2.5
Fvwm is a window manager for X, which provides a virtual/multiple
disjoint desktop, a 3D look for windows decorations, shaped/color
icons. It gives a very good emulation of mwm. A nice button-bar
can be used to provide convenient access to frequently used
functions of programs.

Flavors:
	rplay	- build with audio/rplay and provide sound support
	debug	- turn on debugging messages

WWW: http://www.fvwm.org/

Will co-exist with the fvwm in XF4, with the idea of being on the
bleeding edge, with more extras built in and flavors for choice.
Per discussion with todd@