Update to xpilot-4.4.1
Changes:
- When tuning text options at run time like the password the server
would crash.  This has now been fixed.
- Clients now support the firewall patch.
For this the client has two new options: -clientPortStart and -clientPortEnd.
- A new server option -passwordFileName gives the path to a file which
only contains a definition for the server password.  The advantage
of this is that you can set the file permissions so that others
cannot read its contents.
- Fix -blockFrictionVisible (didn't work)
- Ben Armstrong fixed a bug where ball ownership for all connected balls
would be reset whenever a player would leave.