Initial import of ktimeclock 2.0.1
KTimeclock is a personal timeclock to keep track of how much time is
spent working on various projects and the tasks that they involve.
Simply set up a series of projects, each of which can contain multiple
sub-projects and tasks; select the current task that you're working on;
start the clock; and KTimeclock will track the time you've spent working
on that task.