Import gnome-panel-2.0.11 -- This package contains the GNOME panel which the area on your desktop from which you can run applications and applets, and perform other tasks.
The libpanel-applet library is also in this package. This library allows
one to develop small applications which may be embedded in the panel.
These are called applets. Several applets are also supplied - the
Workspace Switcher, the Window List, the Inbox Monitor, the Clock and
the infamous 'Wanda the Fish'.

From marcm for gnome2