Make glib2 step up and take ownership of *all* the directories it creates, to include the locale directories, gtk-doc stuff, aclocal and pkgconfig.
With the new pkg-* tools, special casing these is no longer necessary.

Note: other ports that do not recursively depend on glib2 and have not
yet been similarly updated will cause warnings when glib2 is deleted if
they have files in any of the newly owned directories.  This is expected,
those ports need to catch up :).