Update to exim 4.41.
This is a major update and configuration files are not compatible.
A conversion script is included and installed in

        ${PREFIX}/share/examples/exim4/convert4r4

New co-maintainer and much of the inital work done by Ilya Voronin.

A decision was made to include the semi-official exiscan patch in
the default installation, as most admins appear to use this. It is
hoped that is will become part of the core exim4 code "soon".

This port now creates an _exim user and group if they don't already
exist.

Work and testing by Ilya Voronin, Richard Welty, Axel Rau and others
who I can't now find in my mailbox. Apologies if I left you out.