Initial import of scons-0.96.1.
SCons is an Open Source software construction tool--that is, a
next-generation build tool. Think of SCons as an improved,
cross-platform substitute for the classic Make utility with integrated
functionality similar to autoconf/automake and compiler caches such as
ccache. In short, SCons is an easier, more reliable and faster way to
build software.

Submitted and maintained by Andrew Dalgleish <openbsd at ajd.net.au>