Import Subversion 1.1.1
Subversion is a free/open-source version control system. That is,
Subversion manages files and directories over time. A tree of files
is placed into a central repository. The repository is much like an
ordinary file server, except that it remembers every change ever made
to your files and directories. This allows you to recover older versions
of your data, or examine the history of how your data changed.
In this regard, many people think of a version control system as a
sort of time machine.

Joint work with msf@, Sigfred H?versen, Alex Holst and Steven Mestdagh

ok naddy@, go ahead msf@