Fix 2 security issues...
A bug exists in the code that parses responses from Gopher servers.
The bug results in a buffer overflow if a Gopher server returns a
line longer than 4096 bytes.  The overflow results in memory
corruption and usually crashes Squid.

CAN-2005-0094

A bug exists in the code that parses WCCP messages.  An attacker
that sends a malformed WCCP messages, with a spoofed source address
matching Squid's "home router" can crash Squid.

CAN-2005-0095