Initial import of p5-MasonX-Interp-WithCallbacks 1.13
This module brings support for a sort of plugin architecture based on
Params::CallbackRequest to Mason. Mason then executes code before
executing any components. This approach allows you to carry out logical
processing of data submitted from a form, to affect the contents of the
Mason request arguments (and thus the %ARGS hash in components), and even
to redirect or abort the request before Mason handles it.