Import mail/mairix
DESCR:

mairix is a program for indexing and searching email messages stored in
Maildir, MH or mbox folders. The search mode populates a "virtual"
maildir (or MH) folder with symlinks which point to the real messages.
This folder can be opened as usual in your mail program.

From: David Cathcart <david@cathcart.cx>

ok robert@