fix a bunch of issues: - make sure locks happen when they should, including for lib-depends-check. - make sure various targets set PACKAGING as they should. - let BULK=Yes be invoked only from make package.
Fix for the BULK_COOKIE and UPDATE_COOKIE issue. They can no longer depend
on PACKAGE_COOKIES, since this needs to get built with PACKAGING set. So
instead, simply remove the cookies when we create a new package succesfully.

By any kind of reasoning, both bulk and update are `reset' when a new
package appears, so this seems to be the right semantics in most cases.

Thanks for wilfried and nikolay for error spotting. Hoping this will be
what's needed, finally.