import p5-File-Rotate-Backup 0.10
This module will make backups and rotate them according to your
specification. It creates a backup directory based on the file_prefix
you specify and the current time. It then copies the directories you
specified in the call to new() to that backup directory. Then a tar'd
and compressed file is created from that directory. By default, bzip2 is
used for compression.

based on submission by Matthew Elmore <matt at mattelmore dot com>