Import tuxpaint-0.9.16
Tux Paint is a free, award-winning drawing program for children ages 3
to 12 (preschool and K-6).  It combines an easy-to-use interface, fun
sound effects, and an encouraging cartoon mascot who guides children as
they use the program.
Kids are presented with a blank canvas and a variety of drawing tools to
help them be creative.

"fine for me" kili@, ok simon@