use the new features of pkg_create. A few packages may no longer build without a packagename bump, or because they have too long comments.
steven@ and I cleaned up most of them.

As a result, there's no longer any WRKPKG directory with their temporary
files.

We also use the `sort -u' feature of pkg_create for dependencies, so that
the command lines to pkg_create get simpler, which will allow for easier
error-checking later on.