Import gutenprint-5.0.1
Gutenprint, formerly known as gimp-print, is a project to provide high
quality drivers for Canon, Epson, Lexmark and PCL printers.
This package contains the driver libraries which are the basis of the
gutenprint project, and which can be used by other printing programs. It
may be used with most common UNIX print spooling systems, including
CUPS, lpr, LPRng, or others.

There are 3 subpackages:
* foomatic-db-gutenprint -> Foomatic data for gutenprint drivers
* gutenprint-cups -> gutenprint CUPS driver
* gutenprint-gimp -> Gimp print plug-in


Note: the direct CUPS drivers won't work as they need pstoraster which
will be included with the next ghostscript update; you will still be
able to print with CUPS using foomatic (which most of the time is the
recommended way anyway).
Hopefully, more people will test this now that it's in.

based on a port by jakemsr@
ok jakemsr@