import gerbv 1.0.2
Gerber Viewer (gerbv) is a viewer for Gerber files. Gerber files are
generated from PCB CAD system and sent to PCB manufacturers as basis for
the manufacturing process. The standard supported by gerbv is RS-274X.
The basic difference between RS-274D (the old standard) and RS-274X is
basically the addition of apertures in RS-274X. It might be possible to
make an RS-274X file out of an RS-274D file and an aperture list.

gerbv also supports drill files. The format supported are known under
names as NC-drill or Excellon. The format is a bit undefined and
different EDA-vendors implement it different. But basically you need to
have the tools definition in the file, then the parser is quite
tolerant. The different holes are shown as dots in the (scaled) correct
size.

from Andreas Bihlmaier <andreas.bihlmaier at gmx.de> (MAINTAINER), with some
tweaks by me

reminded by claudio@