import ccrtp
GNU ccRTP is an implementation of RTP, the real-time transport protocol
from the IETF (see RFC 3550, RFC 3551 and RFC 3555). ccRTP is a C++
library based on GNU Common C++ which provides a high performance,
flexible and extensible standards-compliant RTP stack with full RTCP
support. The design and implementation of ccRTP make it suitable for
high capacity servers and gateways as well as personal client
applications.

from brad@

ok ajacoutot@