import amoebax
Amoebax is a free multi-platform match-3 puzzle game where the objective
is to beat your opponent in a battle by filling their grid up to the top
with garbage.
You can play as Kim or Tom against six cute creatures controlled by the
amoebas or you can play up to four players in a tournament to check out
who is the amoebas' master.

feedback & ok ajacoutot@