The Horde Team is pleased to announce the final release of the Nag Task List Manager version H3 (2.1.4).  And I am pleased to announce my 1000st committ ;)
This is a security release that fixes privilege escalation in the Horde API
and missing ownership validation in the share management. All users are
encouraged to upgrade to this version.

The major changes compared to the Nag H3 (2.1.3) version are:
    * Fixed privilege escalation in the Horde API.
    * Fixed missing ownership validation on share changes.
    * Updated Japanese, Polish, and Spanish translations.

The full list of changes (from version H3 (2.1.3)) can be viewed here:
http://cvs.horde.org/diff.php/nag/docs/CHANGES?r1=1.115.2.54&r2=1.115.2.56.2.3&ty=h