Import libgringotts-1.2.1
libGringotts is a small, easy-to-use, thread-safe C library originally
developed for Gringotts.  Its purpose is to encapsulate data (generic:
ASCII, but also binary data) in an encrypted and compressed structure,
to be written in a file or used elseway.  It makes use of strong
encryption algorithms, to ensure the data are as safe as possible, and
allow the user to have the complete control over all the algorithms used
in the process.

For encryptions, libGringotts makes use of the MCrypt and MHash libs.
It includes some other useful features, like generation of encrypted
temporary files, secure memory management, and functions for the
evaluation of password quality.


from Pierre-Emmanuel Andre <pea at raveland dot org> (maintainer) with
some tweaks