import gnopernicus 1.1.2
Gnopernicus is part of the GNOME Accessibility Project. It is a screen
reader and magnifier and it enables users with limited vision, or no
vision, to use the Gnome 2 desktop and Gnome/GTK+2 applications
effectively. By providing automated focus tracking and fullscreen
magnification, Gnopernicus aids low-vision Gnome users.

feedback and ok ajacoutot@