Initial import bgs.
bgs allows you to tailor the appearance of the background ("root")
window on a workstation display running X. It uses imlib2 for image
rendering and rotates the images automatically. It is designed for
dynamic Xinerama/Xrandr setups such as those used with notebooks, but it
works well in any setup.

From James Turner (MAINTAINER)
ok okan@