Import xerces-c-2.8.0
Xerces-C++ is a validating XML parser written in a portable subset of
C++.  Xerces-C++ makes it easy to give your application the ability to
read and write XML data.  A shared library is provided for parsing,
generating, manipulating, and validating XML documents using the DOM,
SAX, and SAX2 APIs.  For an introduction to programming with Xerces-C++
refer to the Programming Guide.
<...>

ok landry@