import nn 6.7.3
nn is a menu based (point and shoot) netnews reader with a complete set
of features to satisfy both the expert and the novice user. Since its
first release in Denmark in 1984 (!), in Europe in 1988, and the global
release in June 1989, it has replaced rn and other well-known news
readers at many sites.

from aaron w. hsu (MAINTAINER) with tweaks by ajacoutot@ and me