Import xapian-core-1.0.13
Xapian is an opensource search engine library.
It is a highly adaptable toolkit which allows developers to easily add
advanced indexing and search facilities to their own applications. It
supports the Probabilistic Information Retrieval model and also supports
a rich set of boolean query operators.

This is a prerequisite of porting pinot.