Import devel/fossil, from James Turner with a couple of tweaks by me;
Fossil is a software configuration management system. Fossil is software
that is designed to control and track the development of a software project
and to record the history of the project. There are many such systems in
use today. Fossil strives to distinguish itself from the others by being
extremely simple to setup and operate.

In addition to distributed version control, it supports distributed bug
tracking and wiki, with a built-in web interface, in a single easy-to-use
integrated package.