import libgdata 0.4.0
libgdata is a GLib-based library for accessing online service APIs using
the GData protocol - most notably, Google's services. It provides
APIs to access the common Google services, and has full asynchronous
support.

this used to be part of evolution-data-server, but from e-d-s 2.28 onwards
it's split into it's own library. (nothing picks it up for now, tested in a
bulk build)

ok ajacoutot@