import ports/databases/pkglocatedb:
Database of files from all packages in the ports tree, including names
of the packages containing them. This is in a format suitable for use with
locate(1) (either with -d, or in a colon-separated list in LOCATE_PATH).

Note: like sqlports, this is generated from the ports tree itself,
so if your tree is in bad shape re: conflicts etc, this won't build.

ok espie, discussed with many, reminded by oga@