Import net/py-xmpppy Python XMPP and Jabber implementation
xmpppy is a Python library that is targeted to provide easy scripting
with Jabber. Similar projects are Twisted Words and jabber.py.

This library was not designed from scratch. It inherits some code from
jabberpy and have very similar API in many places. Though it is separate
project since it have almost completely different architecture and
primarily aims to work with jabberd2 - the new Open Source Jabber Server.

OK, landry@