Import gnome-main-menu-0.9.15.
The gnome-main-menu is a set of three utilities.  A gnome-panel applet
similar to the traditional main-menu but a few addtions and changes.
In particular the gnome-main-menu does not provide direct access to all
the applications on the system, but rather displays user specified
"favorites" and recently launched applications.

<snip>

This is the default GNOME menu that SUSE Linux uses.
ok jasper@