import hs-blaze-builder 0.2.1.1
This library provides an abstraction of buffered output of byte streams
and several convenience functions to exploit it. For example, it allows
to efficiently serialize Haskell values to lazy bytestrings with a large
average chunk size. The large average chunk size allows to make good use
of cache prefetching in later processing steps (e.g. compression) and
reduces the sytem call overhead when writing the resulting lazy
bytestring to a file or sending it over the network.

ok kili@