Import ports/net/icinga (with -I !, so the 'core' directory doesn't become cvs-food). See the README file for information about migrating from Nagios with minimal config changes.
The icinga-web API-backed user interface is a work in progress
and not yet linked to the build; the core program and classic web
interface are perfectly usable.

Comments/ok from pea@, thanks for testing!

-- -- --
Icinga monitors network resources such as servers, routers, switches,
applications and services and notifies administrators or responsible
persons. It's a fork of Nagios and supports its extensive pool of
monitoring plug-ins. Current status information, historical logs
and reports can all be accessed via a web browser.
-- -- --